refactor(conditions): refactor conditon related objects into modules
