+++ packages/taco/src/conditions/context/context.ts
@@ -2,7 +2,7 @@ import { Context, Conditions as WASMConditions } from '@nucypher/nucypher-core';
 import { fromJSON, toJSON } from '@nucypher/shared';
 import { ethers } from 'ethers';
 
-import { CompoundConditionType } from '../base/compound-condition';
+import { CompoundConditionType } from '../compound-condition';
 import { Condition, ConditionProps } from '../condition';
 import { ConditionExpression } from '../condition-expr';
 import {
@@ -180,11 +180,9 @@ export class ConditionContext {
     signer?: ethers.Signer,
     customParameters?: Record<string, CustomContextParam>,
   ): ConditionContext {
-    const innerCondition =
-      ConditionExpression.fromWASMConditions(conditions).condition;
     return new ConditionContext(
       provider,
-      innerCondition,
+      ConditionExpression.fromWASMConditions(conditions).condition,
       customParameters,
       signer,
     );
