+++ packages/taco/src/conditions/predefined/index.ts
@@ -1,1 +1,1 @@
-export { ERC721Balance, ERC721Ownership } from './erc721';
+export * as erc721 from './erc721';
