+++ packages/taco/test/conditions/base/condition.test.ts
@@ -1,9 +1,9 @@
 import { TEST_CONTRACT_ADDR } from '@nucypher/test-utils';
 import { describe, expect, it } from 'vitest';
 
-import { Condition } from '../../../src/conditions';
 import { ContractCondition } from '../../../src/conditions/base';
-import { ERC721Ownership } from '../../../src/conditions/predefined';
+import { Condition } from '../../../src/conditions/condition';
+import { ERC721Ownership } from '../../../src/conditions/predefined/erc721';
 import { fakeCondition, testContractConditionObj } from '../../test-utils';
 
 describe('validation', () => {
