+++ packages/taco/test/conditions/base/contract.test.ts
@@ -2,10 +2,6 @@ import { initialize } from '@nucypher/nucypher-core';
 import { fakeProvider, fakeSigner } from '@nucypher/test-utils';
 import { beforeAll, describe, expect, it } from 'vitest';
 
-import {
-  ConditionExpression,
-  CustomContextParam,
-} from '../../../src/conditions';
 import { ContractCondition } from '../../../src/conditions/base';
 import {
   ContractConditionProps,
@@ -13,7 +9,9 @@ import {
   ContractConditionType,
   FunctionAbiProps,
 } from '../../../src/conditions/base/contract';
+import { ConditionExpression } from '../../../src/conditions/condition-expr';
 import { USER_ADDRESS_PARAM } from '../../../src/conditions/const';
+import { CustomContextParam } from '../../../src/conditions/context';
 import { testContractConditionObj, testFunctionAbi } from '../../test-utils';
 
 describe('validation', () => {
