+++ packages/taco/test/conditions/context.test.ts
@@ -4,16 +4,21 @@ import { ethers } from 'ethers';
 import { beforeAll, describe, expect, it } from 'vitest';
 
 import { toBytes, toHexString } from '../../src';
-import { ConditionExpression, CustomContextParam } from '../../src/conditions';
 import {
   ContractCondition,
   ContractConditionProps,
-  ReturnValueTestProps,
   RpcCondition,
 } from '../../src/conditions/base';
-import { paramOrContextParamSchema } from '../../src/conditions/base/shared';
+import {
+  paramOrContextParamSchema,
+  ReturnValueTestProps,
+} from '../../src/conditions/base/shared';
+import { ConditionExpression } from '../../src/conditions/condition-expr';
 import { USER_ADDRESS_PARAM } from '../../src/conditions/const';
-import { RESERVED_CONTEXT_PARAMS } from '../../src/conditions/context/context';
+import {
+  CustomContextParam,
+  RESERVED_CONTEXT_PARAMS,
+} from '../../src/conditions/context/context';
 import {
   testContractConditionObj,
   testFunctionAbi,
